Implement touch-based dragging (no mouse yet)
Follow only the first touch.
Disable click events when the touch is near/in the field.

The mapping of touch points to field co-ordinates is a little jank at
the moment using a hardcoded field position. I'm not sure exactly how
we want to do this "for reals"; we'll probably want to "fix" the field
position/size from outside the field and pass it down to centralize the
information. The field should probably only need to render itself, not
calculate how big it should be and where.