Implement basic line clear functionality
We're not checking for win/lose conditions yet, but it's now actually
somewhat "playable". This gives a good idea of how the control scheme
actually feels.

We're still also allowing pieces to be dropped anywhere there isn't
a collision. Which feels annoying when you drop right above the gap
you're trying to fill.

Observations:
- I immediately want to be able to drag from the orientation selectors
  onto the field, not tap then tap-and-drag.

- I also feel like I want to hold the piece in place and change the
  orientation with another finger without letting go.