WIP - Implement multi-touch + (mouse controls)
The Touch API doesn't give us the element currently being touched, the
`target` doesn't change as we move over different elements. So we can't
bind a touchstart/move listener to every element and have it pass it's
own data to the handler. At the moment we've got some duplicate / hard
coded logic that figures out where the touch was and which property
to set.