Consolidate touch logic and field placement
I'm not sure if this is actually a good way to do this; it feels a
little strange in that the `App` has it's own idea of what counts
as interacting with the components. This is not very modular; the
components can't necessarily be replaced/updated to lay things out
differently (internally) without also requiring the interaction
logic in the `App` to change. That said, at this point, that kind
of change doesn't feel very likely (but we'll see).

Another alternative would be to put the currently touched point(s)
into the state and get the components themselves to figure out if a
given canvas point is inside their subcomponents and fire a callback via
prop functions accordingly, which can then set the actual state we want.