Make the active position highlight more subtle
If we keep this in the "final" game, we'll probably want to refactor
the way the field/blocks are handled to skip re-rendering the entire
field whenever the active position changes. Redux can probably help
here e.g. by making the Block a connected component and checking if
it's row / column matches the active position directly, without the
field having to tell it.