WIP - Fold path check into placeability check
This is probably not a good place for this check for optimisation
purposes.

This is possibly a bit slow to calculate every frame (at least the way
it is currently written); even if it isn't, it's highly wasteful.

In theory we only need to check every possible destination once per
starting piece and field, validity only changes when the piece/field
change, current position/orientation shouldn't make a difference
(provided the validity check is complete/correct).